Rejig code into calculator.js and add row highlighting functionality
