+++ site/src/index.njk
@@ -52,7 +52,7 @@ title: Example
     
     <p class="overview"></p>
     
-    <table class="hidden">
+    <table class="hiden">
       <thead>
         <tr>
           <th>Day</th>
@@ -67,7 +67,6 @@ title: Example
   <!--
   FEATURES:
    - Allow user to highlight any table row with a little + that appears on hover
-   - automatic percentage that on level 2, changes from 2.5% to 2.8% from 501 earnings 
   -->
 </div>
