+++ site/src/style/abstracts/_variables.scss
@@ -5,6 +5,7 @@ $gold: #fdc763;
 $coral: #FF7F50;
 $grey: #B2BEB5;
 $green: #4CBB17;
+$highlight: #FCF0DD;
 
 $fs-0: 3.75rem; //60px
 $fs-1: 2.5rem; //40px
